==> R/04_describe.qmd <==
---
title: "04_describe"
author: "Group_23"
format:
  html:
    embed-resources: true
editor: visual
---

# 04 Describe data

```{r}
#| eval: true 
#| message: false
#| warning: false
library(tidyverse)
```

## 04.1. Load data from .tsv files

```{r}
prot_clin_cl <- read.delim("../data/03_dat_aug.tsv") |>
  as_tibble()
gene_metadata <- read.delim("../data/01_load_gene_metadata.tsv") |>
  as_tibble()
```

## 04.2. Descriptive statistics

In this segment we will focus on the main characteristics of each attribute, such as number of patients per group, mean median, standard deviation, variance, range, and quantiles.

As a remark before starting this analysis and as we mentioned in the cleaning, we kept NAs for the clinical data of the 3 healthy patients, since these samples will be useful for our analysis and we can not drop them, but filling these data with invented or random values would be misleading. Hence, in most of the statistical analysis that follows 3 NA values will appear for the clinical attributes.

### 04.2.0 Dimensions

```{r}
dim(prot_clin_cl)
```

The final data set contains 83 observations and 9 351 variables.

### 04.2.1. Complete TCGA ID

The first column contains the IDs for the 80 patients, 77 with breast cancer and 3 healthy. 3 IDs were duplicated because biological replicates were analyzed in order to check the reliability of the method. Thus, in these 3 cases the `Complete_TCGA_ID` is followed by a `.1` in the repeated sample. As discussed before, these 3 samples did not have clinical data, but since they come from the patient with the same `Complete_TCGA_ID,` we just copied the corresponding data into the `.1` samples.

### 04.2.2. Gender distribution

First, we will take a look at the number of male and female samples:

```{r}
prot_clin_cl |>
  count(Gender)
```

As expected, a great majority of samples was taken from female patients, since according to the World Health Organization (WHO), only approximately 0.5-1% of breast cancers occur in men (data from the 12th of July 2023). 3 of the samples are not classified either as male or female (but this is not relevant as these samples are from the healthy ones).

### 04.2.3. Age at initial pathological diagnosis distribution

Age distribution was previously discussed in the `03_augmentation.qmd` file, were samples were grouped into age ranges, but in this section we will focus more on the mean and other statistically relevant data.

```{r}
prot_clin_cl |>
  summarise(na_count = sum(is.na(Age.at.Initial.Pathologic.Diagnosis)),
            mean = mean(as.numeric(Age.at.Initial.Pathologic.Diagnosis), 
                        na.rm = TRUE),
            median = median(as.numeric(Age.at.Initial.Pathologic.Diagnosis), 
                            a.rm = TRUE),
            sd = sd(as.numeric(Age.at.Initial.Pathologic.Diagnosis), 
                    na.rm = TRUE),
            q25 = quantile(as.numeric(Age.at.Initial.Pathologic.Diagnosis), 
                           0.25, 
                           na.rm = TRUE),
            q75 = quantile(as.numeric(Age.at.Initial.Pathologic.Diagnosis), 
                           0.75, 
                           na.rm = TRUE))
```

From these statistics, we can infer that the mean age at which breast cancer is diagnosed is 59, with most of the cases being diagnosed between the age of 49 and the age of 67. This matches the information provided by the WHO, that estimates that most of the cases, and hence the risk group for breast cancer, is women over 40 years old. The American Cancer Society establishes the median age at the time of breast cancer diagnosis in 62, and they claim that a very small number of women diagnosed with breast cancer are younger than 45. The distribution in our data is similar, being the median age at the time of breast cancer diagnosis 59 for the 83 samples in our data (3 NAs for age of the patient), close to the 62 provided by The American Cancer Society.

### 04.2.4. Age at initial pathological diagnosis grouping

In the data augmentation section, the samples were grouped together by age, into 6 age groups (gaps of 10 years). The age distribution among samples is shown in the histogram below. There are individuals for all intervals, even if some contain far more samples than others.

```{r}
prot_clin_cl |> 
  filter(Age.Ini.Diagnostic.group != 'NA') |>
  ggplot(mapping = aes(x = Age.Ini.Diagnostic.group)) +
  geom_bar(fill = "cornflowerblue", 
           color = "black", 
           alpha = 0.7) +
  labs(title = "Age Distribution Across Samples",
       x = "Age Group",
       y = "Count") +
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5)) +
  scale_x_discrete()
```

### 04.2.4. Menopausal age grouping

This type of grouping is a variation of the previous one, in this case taking into account the risk age factors for the different types of breast cancer. Hence, we have a group called "young patients" for women under the median age of diagnosis for breast cancer, but that are more prompt to develop a TNBC type of tumor. From 45 to 55 years old, we have the age group where breast cancer is normally diagnosed but still contains data from patients younger than the median age where menopause starts (55), so these tumors are less prompt to have the hormone receptors. Lastly, we have the group of the oldest patients from 55 up to 90 years old, where normally women already have gone through menopause and hence the observed tumors are more prompt to have hormone receptors.

We investigated, whether the increased risk of developing a TNBC tumor at a younger age can be elucidated from our data. For this reason, we analyzed the prevalence of TNBC across the three menopausal age groups. To mitigate the impact of one age group's samples dominating the entire data set and being more pronounced than samples belonging to other age groups, we performed normalization before the plotting step.

```{r}
meno_age_group_members <- prot_clin_cl |>
  count(Age.Menopausal.group, 
        name = "Total")

meno_age_group_tnbc_positive <- prot_clin_cl |>
  filter(TNBC == "1") |>
  count(Age.Menopausal.group, 
        name = "TNBC_Positive")

meno_age_group_data <- left_join(meno_age_group_members, 
                                 y = meno_age_group_tnbc_positive, 
                                 by = "Age.Menopausal.group")

meno_age_group_data <- meno_age_group_data %>%
  mutate(Proportion_TNBC_Positive = TNBC_Positive / Total)

meno_age_group_data
```

```{r}
meno_age_group_data |>
  filter(!is.na(Age.Menopausal.group)) |>
  ggplot(mapping = aes(x = Age.Menopausal.group, 
                       y = Proportion_TNBC_Positive, 
                       fill = Age.Menopausal.group)) +
  geom_bar(stat = "identity", 
           color = "black", 
           alpha = 0.7) +
  labs(title = "Normalized Triple-Negative Breast Cancer (TNBC) Across Age Groups",
       x = "\nMenopausal Age Group",
       y = "Proportion",
       fill = "Menopausal Age Group Range") +
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5),
        legend.position = "bottom") +
  scale_x_discrete(labels = c("Pre-menopausal",
                             "Menopausal",
                             "Post-menopausal")) +
  scale_fill_manual(values = c("darkslategray2",
                               "cornflowerblue",
                               "darkslateblue"))

```

```{r}
#| eval: true
#| echo: false
ggsave("../results/normalized_TNBC_across_menopausal.png", 
       plot = last_plot(), 
       device = "png", 
       bg="white")
```

### 04.2.5. ER and PR status

These two columns make reference to the presence or absence of receptors that attach to the hormones estrogen and progesterone, and hence the tumorous cells need these hormones to grow.

-   **ER-positive:** Breast cancers that have estrogen receptors are called ER-positive (or **ER+**) cancers.

-   **PR-positive**: Breast cancers with progesterone receptors are called PR-positive (or **PR+**) cancers.

Knowing the hormone receptor status of the cancer is important for choosing the adequate treatment: if the cancer has one or both of these hormone receptors, hormone therapy drugs can be used to either lower estrogen levels or stop estrogen from acting on breast cancer cells. About 3 in 4 breast cancers have at least one of these receptors. This percentage is higher in older women than in younger women, since hormone receptor-negative cancers are more common in women who have not yet gone through menopause. DCIS should also be checked for hormone receptors.

```{r}
prot_clin_cl |>
  count(ER, 
        PR) 
```

Out of the 80 samples with the clinical data recorded, 26 patients are negative for both receptors (hormone receptor-negative), 10 patients are positive for the estrogen receptor, and 44 patients are positive for both receptors. In our data there are no samples only positive for the PR receptor.

We can then investigate how does the ER- and PR- profile look like across different age groups.

```{r}
prot_clin_cl_er_pr <- prot_clin_cl |>
  group_by(Age.Ini.Diagnostic.group, 
           ER, 
           PR) |>
  count() 

prot_clin_cl_er_pr 
```

```{r}
prot_clin_cl_er_pr |>
      filter(Age.Ini.Diagnostic.group != 'NA')|>
  ggplot(mapping = aes(x = Age.Ini.Diagnostic.group,
                       y = n,
                       fill = interaction(ER, PR))) +
  geom_bar(stat = "identity",
          color = "black", 
          alpha = 0.7) +
  labs(title = "ER and PR Combinations Across Age Groups",
       x = "Age Group",
       y = "Count",
       fill = "ER/PR Combination") +
  scale_fill_manual(values = c("cornflowerblue", 
                               "deeppink3", 
                               "darkgoldenrod2")) +
  theme_minimal() +
  theme(plot.title = element_text(hjust = 0.5),
        legend.position = "bottom")
```

```{r}
#| eval: true
#| echo: false
ggsave("../results/ER_PR_across_age_groups.png", 
       plot = last_plot(), 
       device = "png", 
       bg="white")
```

### 04.2.6. HER2 status

According to The American Cancer Society, about 15% to 20% of breast tumors have higher levels of a protein known as HER2 (human epidermal growth factor receptor 2 ). These cancers are called HER2-positive breast cancers and tend to grow and spread faster than breast cancers that are HER2-negative, but are much more likely to respond to treatment with drugs that target the HER2 protein.

```{r}
prot_clin_cl |>
  count(HER2)
```

Out of the 80 samples with this clinical data recorded (3 samples have NA values for this attribute), 60 patients are HER-2 negative, 19 HER-2 positive and 1 is equivocal. The "equivocal" status corresponds to a result of 2+ in the immunohistochemistry (IHC) test, which is one of the tests used to find out if cancer cells have a high level of the HER2 protein. It means that the HER2 status needs to be tested with a fluorescence *in situ* hybridization (FISH) test to clarify the result.

### 04.2.7. Triple-negative breast cancer (TNBC) case count

Triple-negative breast tumors don't have too much HER2 and also don't have estrogen or progesterone receptors. Hormone therapy and drugs that target HER2 are not helpful in treating these cancers, but chemotherapy is still useful. These cancers tend to be more common in women younger than 40 years of age, and those who are African descent, or who have a mutation in the *BRCA1* gene. Triple-negative breast cancers grow and spread faster than most other types of breast cancer. Furthermore, even if TNBC tumor cells are sensitive to chemotherapy, early relapse and visceral metastasis are common in this type of breast cancer.

```{r}
prot_clin_cl |>
  count(TNBC) 
```

From this table we can extract that 19 patients are TNBC (encoded as 1 in this variable). The 61 remaining patients (83 total minus 3 NA samples for these attributes) show either one of the receptors or high HER-2 levels (encoded as 0).

### 04.2.8. ER/PR/HER2 status distribution

This column was created during the augmentation following the encoding:

-   0 : TNBC

-   1: HER2

-   2 : PR

-   3 : ER

-   4 : ER/PR

-   5 : ER/HER2

-   6 : PR/HER2

-   7 : ER/PR/HER2

-   "\*": ER/PR but equivocal results on the test for HER2

The "\*" classification was created because of the observation were the immunohistochemistry (IHC) test resulted in a 2+ value, which means "equivocal" and further testing is needed, so we expect only one observation inside this classification.

```{r}
prot_clin_cl |>
  count(ER_PR_HER2)
```

The values above follow what was previously discussed: there are 19 TNBC patients, 7 positive for high levels of HER-2, 6 only positive for the estrogen, 35 positive for both hormone receptors, 5 positive for the combination of estrogen receptor and HER2, and 7 patients positive for all three test, besides the patient for which both ER and PR test were positive, but the test for the HER2 levels was inconclusive. It is worth highlighting that there are no patients with only progesterone receptors or with a combination of progesterone and high HER-2 levels, hence these progesterone receptors are less common in our data set.

We can also see these types of cancer distributed by age:

```{r}
prot_clin_cl |>
  group_by(Age.Menopausal.group, 
           ER_PR_HER2) |>
  count()
```

As pointed out before, the most common cancer among the younger patients is the type 0 or TNBC, while both in the meddle-aged and older groups of patients, the most common cancer is type 4 or ER/PR positive.

### 04.2.9. Tumor stage distribution

The variable "Tumor" refers to a classification ranging from 1 to 4, accounting for the size and location of the tumor. According to the National Cancer Institute (NCI):

-   The T refers to the size and extent of the main tumor. The main tumor is usually called the primary tumor.

-   T1, T2, T3, T4: Refers to the size and/or extent of the main tumor. The higher the number after the T, the larger the tumor or the more it has grown into nearby tissues.

```{r}
prot_clin_cl |>
  count(Tumor)
```

Most of the tumors in our data set are stage 2, followed by 11 cases both for stage 1 and stage 2 tumors.

```{r}
prot_clin_cl |>
  group_by(Age.Ini.Diagnostic.group, 
           Tumor) |>
  count() 
```

When accounting for both the age range and the stage of the tumor, the stage 2 is the most common across most age groups, with the exception of patients between 70 and 80 years-old, where the most common stage is T3. It is also worth mentioning that T4 tumors are rarely diagnosed before 60 years-old, and we can only find T1 and T2 tumors in the youngest rage of patients.

### 04.2.9. Node distribution

The variable "Node" refers to the number of nearby lymph nodes that are affected by cancer. According to the National Cancer Institute:

-   N0: There is no cancer in nearby lymph nodes.

-   N1, N2, N3: Refers to the number and location of lymph nodes that contain cancer. The higher the number after the N, the more lymph nodes that contain cancer.

```{r}
prot_clin_cl |>
  count(Node)
```

Most of the patients in our data set do not have cancer in the nearby lymph nodes. This is relevant as the spread of cancer to the lymph nodes can be an indicator of metastasis or the stage of the cancer.

### 04.2.10. Node Cancer

This variable condenses the information from the variable before, classifying as 1, which means Positive, every tumor with regional lymph nodes (N) and as 0, which means Negative, the N0 tumors. With the data extracted before, we can infer that there are 40 negative samples, 40 positive, and 3 without clinical data for this variable (the healthy samples).

```{r}
prot_clin_cl |>
  group_by(Age.Menopausal.group, 
           Node_Cancer) |>
  count() 
```

We can see that the presence of nearby lymph nodes that have cancer does not seem to be dependent on the age, since all three age groups apparently have the same distribution of "Positives" and "Negatives" (1s and 0s) for this variable.

### 04.2.11. Metastasis

The "Metastasis" variable (M) indicates whether cancer has spread to other parts of the patient's body than its original site.

```{r}
prot_clin_cl |>
  count(Metastasis)
```

Out of the 80 patients with cancer in this data set, only one has metastasis.

### 04.2.12. AJCC (simplified)

As explained in the augmentation file, this column simplifies the information previously encoded in the `AJCC.Stage` and `Converted.Stage` columns.

The American Joint Committee on Cancer (AJCC) for breast cancer is a staging system for classifying the extent of a cancer, based on the size of the tumor (T), lymph node involvement (N), and the presence or absence of metastasis (M). The most recent update to the staging system integrates molecular markers with disease extent for more optimal estimation of prognosis.

Stages 1-4 are described with the Roman numerals I, II, III, and IV. Sometimes these stages are further subdivided using the letters A, B and C that indicate certain characteristics found on work-up of the cancer.

-   **Stage 0** -- This is precancerous and non-invasive. This is also known as "carcinoma *in situ*", or cancer that is only present in the cells where it started. It has not spread to any nearby tissues.

-   **Stage I** -- The tumor is usually small and has just begun growing into nearby tissues, but its invasion does not have much depth. Stage 1 tumors have not spread to nearby lymph nodes or major organs.

-   **Stage II** -- These tumors are larger, have grown into nearby tissue, and have potentially spread into the lymph nodes.

-   **Stage III** -- These cancers are more advanced and aggressive than Stage II.

-   **Stage IV** -- This is the highest stage. These cancers have metastasized, or spread to distant parts of the body.

```{r}
prot_clin_cl |>
  count(AJCC.Simp)
```

In our data, most of the tumors are Stage II according to the AJCC classification, followed by Stage III, I and finally IV with only one case. If we observe the distribution by age:

```{r}
prot_clin_cl |>
  group_by(Age.Menopausal.group, 
           AJCC.Simp) |>
  count() 
```

In this table we can see that it is the oldest age group, where Stage III is found to be the most prevalent and where the only case of Stage IV cancer is observed.

### 04.2.13. Vital state distribution

This variable encodes the deceased (1) and alive (0) patients.

```{r}
prot_clin_cl |>
  count(Vital.State)
```

Seven patients in this data set died. We can also see the distribution by age:\

```{r}
prot_clin_cl |>
  group_by(Age.Menopausal.group, 
           Vital.State) |>
  count() 
```

With this we can see that the deceased patients belonged to the group with the oldest patients (6 out of 7), while one of the deceased patients belonged to the youngest group, where the TNBC has a higher incidence.

It can also be interesting to see the distribution by cancer stage:

```{r}
prot_clin_cl |>
  group_by(AJCC.Simp, 
           Vital.State) |>
  count()
```

Where we can see that most of the deceased patients had a Stage II tumor, which should not be the case for stage II tumors with early treatment. If we combine both age and stage to see the distribution of deceased, we get:

```{r}
prot_clin_cl |>
  filter(Vital.State==1) |>
  group_by(AJCC.Simp, 
           Age.Menopausal.group) |>
  count()
```

We can see that most of the Stage II deceased patients are in the oldest age group, which indicates that age is also an important factor in the 7 registered deaths.

### 04.2.14. PAM50.mRNA distribution

According to the Nurses' Health Study Cohorts, PAM50 is a 50-gene signature that classifies breast cancer into five molecular intrinsic subtypes: Luminal A, Luminal B, HER2-enriched, Basal-like, and Normal-like. Each of the five molecular subtypes varies by their biological properties and prognoses. Luminal A generally has the best prognosis; HER2-enriched and Basal-like are considered more aggressive diseases.

```{r}
prot_clin_cl |>
  count(PAM50.mRNA)
```

Luminal B and Luminal A are the most common types of cancer in our data, while HER2-enriched and Basal-like, are slightly less common.

```{r}
prot_clin_cl |>
  group_by(Age.Menopausal.group, 
           PAM50.mRNA) |>
  count()
```

Both in the youngest and oldest age groups, the most common type of cancer is Basal-like, which is one of the most aggressive, but for the middle-aged group the most common is Luminal A, the one generally with the best prognosis.

After the analysis of all this variables, we can infer that age is an important factor in the development and prognosis of breast cancer, since most of the analyzed variables have a distribution determined by age.

### 04.2.15. Protein expression data

The remaining columns contain protein expression data. They contain expression values for \~12 000 proteins for each sample, with missing values removed.

Log2 iTRAQ ratios were used to represent the relative abundance of proteins in the samples. iTRAQ (Isotope-Tagged Retention And Capture) is a quantitative proteomics technique that measures the abundance of proteins in a sample. Then, log2 iTRAQ ratios were calculated by taking the logarithm base 2 of the ratio between the heavy and light isotope pairs.

The code below provides the mean and the standard deviation of each of the protein columns:

```{r}
prot_clin_cl |>
  select(starts_with("NP"),
         starts_with("XP"),
         starts_with("YP")) |>
  summarize(across(everything(), 
                   list(
                     mean = mean, 
                     sd = sd),
                   .names = "{col}-{fn}")) |>
  pivot_longer(cols = everything(), 
               names_to = "variable", 
               values_to = "value") |>
  separate(variable, 
           into = c("Protein", 
                              "stat"), 
           sep = "-") |>
  pivot_wider(names_from = stat, 
              values_from = value)
```
